# identity-ms/identity-ms-api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "1.2.0-SNAPSHOT",
    "title" : "WSO2UserManagerAPI"
  },
  "basePath" : "/",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/admin/users/roles" : {
      "put" : {
        "tags" : [ "User API" ],
        "summary" : "Assign one or more roles to user",
        "description" : "",
        "operationId" : "updateRoleListOfUser",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "username",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Update User"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : ""
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/roles" : {
      "get" : {
        "tags" : [ "Role API" ],
        "summary" : "Get a list of all the roles created in the system",
        "description" : "",
        "operationId" : "getRoles",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "Valid Roles are found",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          },
          "404" : {
            "description" : "Valid Roles are not found"
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    },
...
    "/users/roles" : {
      "get" : {
        "tags" : [ "User API" ],
        "summary" : "Get the list of roles that a particular user belongs to",
        "description" : "",
        "operationId" : "getRolesOfUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "username",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Valid user is updated",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string"
              }
            }
          }
        },
        "security" : [ {
          "X-Auth-Token" : [ ]
        } ]
      }
    }
  },
  "securityDefinitions" : {
    "X-Auth-Token" : {
      "type" : "apiKey",
      "name" : "X-Auth-Token",
      "in" : "header"
    }
  },
...
}